Logging (#5)
* Add IntelliJ .idea to gitignore

* Add barebones version of logging

Create two new Mongoose models and use them within `updateProject` in the project controller. Add the participants JSON and a utility for checking if a user is in it.

* Use autosave for logging

Remove the ability to configure autosave, so that it is now permanently on for every user. Add an `autosaveEvery` constant for configuring save interval.

* Log on run

Add a new action, controller, and endpoint for logging project runs. Add a new "run" type to the LogItem model.

* Minor fixes

* Allow for logging unsaved projects by saving them first

Refactor logging code into a new function within project.controller.js and add more consistent permission checks.

* Fix small bug with undefined github token

* Differentiate between auto and manual runs

Also fix a bug with storing "run" log files.

Co-authored-by: Anton <antonoutkine@gmail.com>