Study participants table (#47)
* Evaluate study participation based on Mongo table

Add a mongoose model and change the `isPartOfStudy` script to query it. Add a stub file to automatically insert study participation for the developers.

* Prevent account generation when a user is not in the whitelist

* Rename "whitelist" to "allowlist"

* Fix login errors not showing on the frontend

The backend signals a login error through a query param. Under the previous codebase, this query param was lost after the frontend redirected the user from the '/account' page to the '/login' page, since a user logging in would never be logged in, and so would not have the ability to view the '/account' page. Fix this and display an updated error.